There seems to be an issue with (at least) anonymous unions that have the 'transparent_union' attribute.  The symptom is:
1) a type for such a union is generated and passed to the debug hook 'ffi_debug_type_decl', and we output it.  This type doesn't seem to have the transparent_union attribute bit set (yet).

2) later, a type declaration that (conceptually) refers to this anonymous type; the anonymous union object which is the target type is not EQ to the one which we've processed (and, in fact, we've never seen the target type object before, though we saw another anonymous union with the same fields.

A crude workaround is check for this case when processing a typedef and emit the new type (the anonymous union) before emitting the deftype; that means that both versions of the anonymous union are written to the .ffi file (one of them referenced and the other not.)

It's not clear that this is an adequate fix: there may be other ways of referencing these types (e.g., via a pointer type or as a record field).  This does seem to fix the one case I saw (involving the
__SOCKADDR_ARG and __CONST_SOCKADDR_ARG types in <sys/socket.h> on Linux.  Really handling that case
might involve other FFI changes not related to the translator.







git-svn-id: http://svn.clozure.com/publicsvn/ffigen4/trunk@44 df28d2ca-d578-4763-9ef7-37c90b4b449d